fixing

+++ client/src/main/java/itmo/lab6/basic/baseclasses/Coordinates.java
@@ -7,11 +7,10 @@ import java.io.Serializable;
  * The Coordinates class represents a pair of coordinates (x, y).
  * It implements the Serializable interface, allowing instances to be written to and read from streams.
  *
- * @author dorlneylon
- * @version 1.0?
- * @since ??.??.????
  */
 public class Coordinates implements Serializable {
+	//@Serial
+	//private static final long serialVersionUID = 6529685098267757690L;
 	private float x;
 	private Double y;
 
@@ -32,7 +31,7 @@ public class Coordinates implements Serializable {
 	 */
 	public Coordinates() {
 		this.x = (float) (Math.random() * 1000 - 521);
-		this.y = (Double) (Math.random() * 256);
+		this.y =  Math.random() * 256;
 	}
 
 	/**
